+++ src/SymbolicAWEModels.jl
@@ -99,8 +99,6 @@ export init_module
 export update_plot_observables!
 export animate
 export load_sys_struct_from_yaml
-export update_yaml_from_sys_struct!
-export update_sys_struct_from_yaml!
 export replay
 export record
 export plot_sphere_trajectory
@@ -150,7 +148,6 @@ function plot_sphere_trajectory end
 function plot_body_frame end
 function plot_aoa end
 function find_steady_state! end
-function update_yaml_from_sys_struct! end
 function make_lin_sys_state end
 function create_model_archive end